+++ documentify-project/documentify-core/src/main/kotlin/io/github/bgmsound/documentify/core/specification/schema/ResourceSpec.kt
@@ -1,8 +1,9 @@
-package io.github.bgmsound.documentify.core.specification
+package io.github.bgmsound.documentify.core.specification.schema
 
 import com.epages.restdocs.apispec.ResourceDocumentation
 import com.epages.restdocs.apispec.ResourceSnippetParameters
 import com.epages.restdocs.apispec.Schema
+import io.github.bgmsound.documentify.core.specification.DocumentableSpec
 import io.github.bgmsound.documentify.core.specification.element.Link
 import io.github.bgmsound.documentify.core.specification.element.field.Field
 import io.github.bgmsound.documentify.core.specification.schema.request.RequestSpec
@@ -28,17 +29,17 @@ class ResourceSpec(
     }
 
     fun link(rel: String): Link {
-        val link = Link.newLink(rel)
+        val link = Link.Companion.newLink(rel)
         this.links.add(link)
         return link
     }
 
     fun links(vararg links: String) {
-        this.links.addAll(links.map { Link.newLink(it) })
+        this.links.addAll(links.map { Link.Companion.newLink(it) })
     }
 
     fun links(links: Collection<String>) {
-        this.links.addAll(links.map { Link.newLink(it) })
+        this.links.addAll(links.map { Link.Companion.newLink(it) })
     }
 
     fun tag(tag: String) {
@@ -70,7 +71,7 @@ class ResourceSpec(
     }
 
     override fun build(): List<Snippet> {
-        val resourceBuilder = ResourceSnippetParameters.builder()
+        val resourceBuilder = ResourceSnippetParameters.Companion.builder()
         if (tags.isNotEmpty()) {
             resourceBuilder.tags(*tags.toTypedArray())
         }
@@ -93,13 +94,13 @@ class ResourceSpec(
         }
         if (request.fields.isNotEmpty()) {
             if (request.schema != null) {
-                resourceBuilder.requestSchema(Schema.schema(request.schema!!))
+                resourceBuilder.requestSchema(Schema.Companion.schema(request.schema!!))
             }
             resourceBuilder.requestFields(buildFields(request.fields))
         }
         if (response.fields.isNotEmpty()) {
             if (response.schema != null) {
-                resourceBuilder.responseSchema(Schema.schema(response.schema!!))
+                resourceBuilder.responseSchema(Schema.Companion.schema(response.schema!!))
             }
             resourceBuilder.responseFields(buildFields(response.fields))
         }
